jb-labeling-service/pyproject.toml
[tool.poetry]
name = "jb-labeling-service"
version = "0.1.0"
description = ""
authors = ["OpenNyAI Team <opennyai@googlegroups.com>"]
readme = "README.md"
packages = [{include = "labeling"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^1.10.7"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
jb-core = {path = "../packages/jb-core", develop = true}
fastapi = "^0.100.0"
asyncpg = "^0.28.0"
jb-auth-service = {path = "../jb-auth-service", develop = true}
jb-auth-token = {path = "../packages/jb-auth-token", develop = true}
openai = "^0.27.8"
tiktoken = "^0.4.0"
pandas = "^2.0.3"
httpx = "^0.24.1"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.3.0"
poethepoet = "^0.20.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
gunicorn = "^20.1.0"
faker = "^18.11.1"
pytest-celery = "^0.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks.build]
shell = """
docker build --platform=linux/amd64  -f tools/Dockerfile -t gcr.io/indian-legal-bert/labeling_service:latest ..
"""
help = "build docker container"

[tool.poe.tasks.test]
cmd = "python -m pytest -vv -o log_cli=1 -o log_cli_level=INFO -W 'ignore::DeprecationWarning' $FILE"
args = [{name="FILE", default="tests", positional=true}]
help = "run tests using pytest"
